worker/Worker/requirements.txt
Django==4.2.16
python-decouple==3.8
celery==5.4.0
boto3==1.35.7
pycurl==7.45.3
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
django-storages==1.14.4
requests==2.32.3
djangorestframework==3.15.2